+++ packages/botonic-plugin-hubtype-babel/src/options.ts
@@ -1,6 +1,0 @@
-export interface PluginHubtypeBabelOptions {
-  projectId: string
-  authToken?: string
-  automaticBotMessagePrefix?: string
-  includeHasSense?: boolean
-}