+++ packages/botonic-react/src/webview-app.tsx
@@ -5,7 +5,11 @@ import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter, Route } from 'react-router-dom'
 
-import { CloseWebviewOptions, WebviewRequestContext } from './contexts'
+import {
+  CloseWebviewOptions,
+  WebviewRequestContext,
+  WebviewRequestContextType,
+} from './contexts'
 
 enum WebviewUrlParams {
   Context = 'context',
@@ -14,21 +18,15 @@ enum WebviewUrlParams {
   HubtypeApiUrl = 'hubtype_api_url',
 }
 
+const DEFAULT_HUBTYPE_API_URL = 'https://api.hubtype.com'
+
 class App extends React.Component {
   private url: URL
-  private botId: string
-  private userId: string
-  private hubtypeApiUrl: string
-  private state: { session: Session; params: Record<string, string> }
+  private hubtypeApiUrl: string = DEFAULT_HUBTYPE_API_URL
 
   constructor(props) {
     super(props)
     this.url = new URL(window.location.href)
-    this.botId = this.url.searchParams.get(WebviewUrlParams.BotId)
-    this.userId = this.url.searchParams.get(WebviewUrlParams.UserId)
-    this.hubtypeApiUrl =
-      this.url.searchParams.get(WebviewUrlParams.HubtypeApiUrl) ||
-      'https://api.hubtype.com'
     this.state = {
       session: null,
       params: {},
@@ -36,39 +34,134 @@ class App extends React.Component {
   }
 
   async componentDidMount() {
+    await this.initializeApp()
+  }
+
+  private async initializeApp() {
     try {
-      const session = await this.getSessionFromUrl()
-      const params = this.getParamsFromUrl()
-      this.setState({ session, params })
+      const botId = this.url.searchParams.get(WebviewUrlParams.BotId)
+      const chatId = this.url.searchParams.get(WebviewUrlParams.UserId)
+      const hubtypeApiUrl = this.url.searchParams.get(
+        WebviewUrlParams.HubtypeApiUrl
+      )
+
+      if (botId && chatId && hubtypeApiUrl) {
+        const session = await this.getBotSessionContextFromExternalApi(
+          hubtypeApiUrl,
+          botId,
+          chatId
+        )
+        this.hubtypeApiUrl = hubtypeApiUrl
+        this.setState({
+          session,
+          params: this.getParamsFromUrl(),
+        })
+      } else {
+        const session = this.getBotSessionContextFromUrl()
+        this.hubtypeApiUrl = session._hubtype_api || DEFAULT_HUBTYPE_API_URL
+        this.setState({
+          session,
+          params: this.getParamsFromUrl(),
+        })
+      }
     } catch (error) {
-      console.error('Error getting bot session context from url', error)
+      console.error('Failed to initialize app:', error)
+      const session = this.getBotSessionContextFromUrl()
+      this.hubtypeApiUrl = session._hubtype_api || DEFAULT_HUBTYPE_API_URL
+      this.setState({
+        session,
+        params: this.getParamsFromUrl(),
+      })
     }
   }
 
-  async getSessionFromUrl() {
-    const url = `${this.hubtypeApiUrl}/external/v2/conversational_apps/${this.botId}/users/${this.userId}/context/`
+  private async getBotSessionContextFromExternalApi(
+    hubtypeApiUrl: string,
+    botId: string,
+    userId: string
+  ) {
+    const url = `${hubtypeApiUrl}/external/v2/conversational_apps/${botId}/users/${userId}/context/`
     const response = await axios.get(url)
     return response.data
   }
 
-  getParamsFromUrl() {
+  private getBotSessionContextFromUrl() {
+    const urlContext = this.url.searchParams.get(WebviewUrlParams.Context)
+    try {
+      return JSON.parse(urlContext || '{}')
+    } catch (error) {
+      console.error('Failed to parse session context from URL:', error)
+      return {}
+    }
+  }
+
+  private getParamsFromUrl() {
     const keysToExclude = [
       WebviewUrlParams.Context,
       WebviewUrlParams.BotId,
       WebviewUrlParams.UserId,
       WebviewUrlParams.HubtypeApiUrl,
     ]
-    const params = Array.from(this.url.searchParams.entries())
-      .filter(([key, _]) => !keysToExclude.includes(key))
-      .reduce((o, [key, value]) => {
-        o[key] = value
-        return o
+    return Array.from(this.url.searchParams.entries())
+      .filter(([key]) => !keysToExclude.includes(key))
+      .reduce((params, [key, value]) => {
+        params[key] = value
+        return params
       }, {})
-    return params
+  }
+
+  private async closeWebviewForProvider(
+    provider: string,
+    session: Session,
+    payload?: string
+  ) {
+    const { user } = session
+
+    switch (provider) {
+      case PROVIDER.WHATSAPP:
+        location.href = `https://wa.me/${user.unformatted_phone_number}`
+        break
+      case PROVIDER.TELEGRAM:
+        location.href = `https://t.me/${user.imp_id}`
+        break
+      case PROVIDER.APPLE:
+        location.href = `https://bcrw.apple.com/urn:biz:${user.imp_id}`
+        break
+      case PROVIDER.TWITTER:
+        location.href = `https://twitter.com/messages/compose?recipient_id=${user.imp_id}`
+        break
+      case PROVIDER.INSTAGRAM:
+        window.close()
+        break
+      case PROVIDER.FACEBOOK:
+        try {
+          window.MessengerExtensions.requestCloseBrowser(
+            () => {}, // success callback
+            () => window.close() // error callback
+          )
+        } catch (error) {
+          window.close()
+        }
+        break
+      case PROVIDER.WEBCHAT:
+        try {
+          await parent.postMessage('botonicCloseWebview', '*')
+        } catch (error) {
+          console.error('Failed to send close message to parent:', error)
+        }
+        break
+      default:
+        console.warn(`Unknown provider: ${provider}`)
+    }
   }
 
   async close(options?: CloseWebviewOptions) {
-    let payload = options ? options.payload : null
+    if (!this.state.session) {
+      console.error('No session available for closing webview')
+      return
+    }
+
+    let payload = options?.payload || null
 
     if (options?.path) {
       payload = `__PATH_PAYLOAD__${options.path}`
@@ -79,59 +172,36 @@ class App extends React.Component {
         payload = `${payload}?${params2queryString(options.params)}`
       }
 
-      const session = this.state.session
       try {
-        const url = `${this.hubtypeApiUrl}/v1/bots/${this.botId}/send_postback/`
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        const data = { payload: payload, chat_id: this.userId }
+        const url = `${this.hubtypeApiUrl}/v1/bots/${this.state.session.bot.id}/send_postback/`
+        const data = {
+          payload,
+          chat_id: this.state.session.user.id,
+        }
         await axios.post(url, data)
-      } catch (e) {
-        console.log(e)
+      } catch (error) {
+        console.error('Failed to send postback:', error)
       }
     }
 
-    const provider = this.state.session.user.provider
-    const impId = this.state.session.user.imp_id
-    if (provider === PROVIDER.WHATSAPP) {
-      const phone_number = this.state.session.user.unformatted_phone_number
-      location.href = 'https://wa.me/' + phone_number
-    }
-    if (provider === PROVIDER.TELEGRAM) {
-      location.href = 'https://t.me/' + impId
-    }
-    if (provider === PROVIDER.APPLE) {
-      location.href = 'https://bcrw.apple.com/urn:biz:' + impId
-    }
-    if (provider === PROVIDER.TWITTER) {
-      location.href =
-        'https://twitter.com/messages/compose?recipient_id=' + impId
-    }
-    if (provider === PROVIDER.INSTAGRAM) {
-      window.close()
-    }
-    if (provider === PROVIDER.FACEBOOK) {
-      try {
-        window.MessengerExtensions.requestCloseBrowser(
-          function success() {},
-          function error(err) {
-            window.close()
-          }
-        )
-      } catch (e) {
-        window.close()
-      }
-    }
-    if (provider === PROVIDER.WEBCHAT) {
-      try {
-        await parent.postMessage('botonicCloseWebview', '*')
-      } catch (e) {}
-    }
+    await this.closeWebviewForProvider(
+      this.state.session.user.provider,
+      this.state.session,
+      payload
+    )
   }
 
   render() {
-    const webviewRequestContext = {
-      session: this.state.session || {},
-      params: this.state.params || {},
+    if (!this.state.session) {
+      return null
+    }
+
+    const webviewRequestContext: WebviewRequestContextType = {
+      params: this.state.params,
+      session: this.state.session,
+      getUserCountry: () => this.state.session.user.country,
+      getUserLocale: () => this.state.session.user.locale,
+      getSystemLocale: () => this.state.session.system_locale,
       closeWebview: this.close.bind(this),
     }
 
@@ -157,8 +227,7 @@ export class WebviewApp {
         <App webviews={this.webviews} locales={this.locales} />
       </BrowserRouter>
     )
-    const container = dest
-    const reactRoot = createRoot(container)
+    const reactRoot = createRoot(dest)
     reactRoot.render(component)
   }
 }