+++ scripts/nlu/txt-to-csv.ts
@@ -1,7 +1,0 @@
-import { CsvGenerator } from './csv-generator'
-
-const srcPath = ''
-const dstPath = ''
-
-const generator = new CsvGenerator()
-generator.generate(srcPath, dstPath)