+++ packages/botonic-plugin-hubtype-analytics/CHANGELOG.md
@@ -10,17 +10,21 @@ All notable changes to Botonic will be documented in this file.
     Click to see more.
   </summary>
   
-## [0.36.0] - 2025-mm-dd
+## [0.36.x] - 2025-mm-dd
 
 ### Added
 
 ### Changed
 
-- [PR-3028](https://github.com/hubtype/botonic/pull/3028): Use new fields user locale, country and system locale.
-
 ### Fixed
 
 </details>
+
+## [0.36.0] - 2025-06-18
+
+### Changed
+
+- [PR-3028](https://github.com/hubtype/botonic/pull/3028): Use new fields user locale, country and system locale.
 
 ## [0.35.0] - 2025-05-06